More changes to the associations objects
The model rules seems to be an exhaustive list of what the association can do. Almost as if - if it's not on the list, we can't do it. So we're trying to cover all bases. But in plain english, not legalese. I ignored the investments part, since I don't think that will be an issue. If it becomes one, we'll have an SGM, and add something to cover it.